+++ parquet/src/file/serialized_reader.rs
@@ -763,9 +763,15 @@ mod tests {
 
     #[test]
     fn test_file_reader_optional_metadata() {
+        // file with optional metadata: bloom filters, encoding stats, column index and offset index.
         let file = get_test_file("data_index_bloom_encoding_stats.parquet");
         let file_reader = Arc::new(SerializedFileReader::new(file).unwrap());
-        let col_metadata = file_reader.metadata.row_group(0).column(0);
+
+        let row_group_metadata = file_reader.metadata.row_group(0);
+        let col0_metadata = row_group_metadata.column(0);
+
+        // test optional bloom filter offset
+        assert_eq!(col0_metadata.bloom_filter_offset().unwrap(), 192);
 
         // test page encoding stats
         assert!(col_metadata.has_page_encoding_stats());